ARM: OMAP2+: powerdomain/PM: only program supported power states
Previously the PM code attempted to program power states that the
underlying powerdomain didn't support.  Change this to only program
power states that are supported by the hardware.

In the future, the PM code should be changed further to program
explicit lists of powerdomain power states that are valid for each
chip (and supported by the underlying software).  Also the powerdomain
code should return errors when PM code tries to program an unsupported
power state, rather than simply ignoring the requrest.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>