ARM: OMAP2+: powerdomain: assume memory bank power states follow powerdomain
Assume that the memory bank power states follow the powerdomain power
states.  The motivations are to reduce the amount of powerdomain code,
decrease the execution time of the powerdomain state switch code, and
simplify the power state debug.

This assumption is true for the currently-implemented functional power
states.  If it ceases to be true, the memory bank power states should
be controlled via the functional power states, rather than controlled
via low-level functions.

For the moment, the underlying implementation code in
mach-omap2/prm*.c has been preserved, although some of that code may
be removable too.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Cc: Tero Kristo <t-kristo@ti.com>
Cc: Rajendra Nayak <rnayak@ti.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>